sec1 crate (#37)
Adds a `sec1` crate similar to the `pkcs1` crate, but for EC keys as
opposed to RSA keys, implementing the SEC1 `EcPrivateKey` ASN.1 DER
schema:

https://www.secg.org/sec1-v2.pdf

This implementation was originally from the `elliptic-curve` crate,
however extracting it here not only makes it reusable but makes possible
a blanet impl of PKCS#8 support for any types which impl
`FromEcPrivateKey`/`ToEcPrivateKey`, similar to the blanket impl
provided for the `pkcs1` crate.

It would eventually probably make sense to extract more of the SEC1
encoding implementation into this crate from the `elliptic-curve` crate,
namely support for the `Octet-String-to-Elliptic-Curve-Point` and
`Elliptic-Curve-Point-to-Octet-String` conversions. However, they are
omitted from an initial implementation for now.