der_derive: fix `#[asn1(type = "...")]` attribute (#204)
PR #202 removed support for the `#[asn1(type = "...")]` attribute when
deriving `Sequence`.

This PR restores it, and also slightly changes the way the `default`
attribute is handled, with added tests.

Additionally, this commit introduces a reference type for `OPTIONAL`
similar to the one for `CONTEXT-SPECIFIC` called `OptionalRef` that can
be used as an `Encodable` trait object for `Option<&T>`. This is used by
the custom derive when encoding fields with defaults.