Only emit `.try_into()?` when a type is specified (#397)
The purpose of this transformation is for when the caller specifies `type =
"..."`. In that case, we decode into the specified type and then convert into
actual type. However, in cases where no type is specified, we have to decode
into an unspecified type `T` and then convert into the final type; but the
compiler (obviously) cannot determine the type `T`.

Therefore, when no `type` is specified, we drop the `.try_into()?` and just
attempt to decode into the final type directly.

Signed-off-by: Nathaniel McCallum <nathaniel@profian.com>