base64ct: bound `Encoding` on `Variant` (#562)
The `Encoding` trait defines this crate's public API, however internally
it's implemented in terms of `Variant`.

`Variant` is a sealed trait. Previously it was exposed as the bound on
the generic parameter of `Decoder` and `Encoder`.

This is technically a breaking change but it theoretically corrects a
deficiency/bug: it was technically possible before for a user of
`base64ct` to impl the `Encoding` trait themselves. But this was never
intended to be supported, and in theory adding this bound on a sealed
trait should both otherwise get the sealed trait out of the public API
as well as ensuring that nothing but `base64ct` itself can impl
`Encoding`.