der: make `Reader::input_len` infallible (#607)
It was previously fallible to support runtime calculations, but with a
bit of refactoring and simplification it's possible to query this for
the current `Decoder` type using `ByteSlice::len`.

Other readers may need to compute it, but this can be performed as an
eager calculation of a fallible constructor (similar to how `ByteSlice`
does it). Since it's a static, non-changing value for the lifecycle of
any given reader it makes sense to make it infallible.

This decision may need to be revisited if we ever want to support
"streaming" readers which don't know the total size of the input data
ahead of time.