Add support for parsing keys out of the ssh known_hosts file format (#624)
This is largely based on the authorized_keys parse functionality,
and there are probably some opportunities to remove redundancy, but
this seems like a good start.

This includes a change to allow empty comments in OpenSSH keys:
Even the comments say this is allowed. This also, however means
that it's not possible to detect a truncated key just from the end
of input, so that case is now treated as the end of the key.

Detection of truncated keys should probably happen at a higher
level, where the correct length of the key is known.