der: add streaming on-the-fly `PemDecoder` and `BitStringOwned` (#636)
Adds a PEM decoder which leverages the `Reader` trait to be able to
decode DER from PEM being decoded on-the-fly. This requires bounding on
`DecodeOwned`, as it's not possible to borrow DER bytes from a PEM
input.

Also adds `BitStringOwned`, an owned form of `BitString`. This may not
be the greatest naming scheme (perhaps we should change the current
`BitString` type to `BitStringRef`).

There are several owned types to add. However this much is capable of
decoding some SPKI test vectors.