der: deprecate `TagNumber::new` (#1727)
After #1651 the inner `u32` of `TagNumber` is now `pub` and the
`TagNumber::new` function no longer maintains an invariant (which it
previously did by panicking when tag numbers are out-of-range).

With the field now `pub` there is no longer a need to call a function to
construct a `TagNumber`, instead it can be done directly.