der: extract `reader::position::Position` (#1880)
Extracts a state management struct for nested position tracking.

The goal is to support nested message parsing keeping track of where we
are in the message using only the call stack, i.e. without the need to
keep such information on the heap, as this crate is intended to support
no_std/no-alloc "heapless" targets.

Extracting this logic into its struct not only makes it potentially
reusable, but makes it significantly easier to test, and this commit
also adds a number of initial tests.